requirements.txt
dj-database-url==0.5.0
Django==2.1
django-appconf==1.0.4
django-bootstrap3==12.1.0
django-config==0.1.4
django-heroku==0.3.1
django-imagekit==4.0.2
gunicorn==20.0.4
pilkit==2.0
Pillow==7.1.2
psycopg2==2.8.5
python-config==0.1.2
python-decouple==3.3
pytz==2020.1
six==1.15.0
whitenoise==5.1.0
